refactor: consolidate CommandExitStatus implementations for local and remote processes
